backend/post/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
PyJWT==2.3.0
requests==2.26.0 